requirement.txt
Django==1.11
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
mysqlclient==1.3.14
PyJWT==1.7.1
pytz==2018.9
